Decompile func0002373c

+++ src/include/game/data/data_000000.h
@@ -145,7 +145,7 @@ extern u32 var8005efcc;
 extern u32 var8005efd0;
 extern u32 var8005efe0;
 extern u32 var8005efec;
-extern u32 var8005f008;
+extern s16 var8005f008;
 extern struct animheader *g_Anims;
 extern u32 var8005f010;
 extern u32 var8005f014;
